Add unit tests for ProtocolAdapterWrapper to validate state transitions and error handling
- Implement tests for initial state, lifecycle methods, and error recovery scenarios.
- Verify correct behavior for starting and stopping adapters, including handling of northbound and southbound connections.
- Ensure proper state transitions and listener notifications during state changes.
- Test concurrent start and stop operations to prevent race conditions.
- Validate shutdown flag behavior to prevent connection status changes during shutdown.
