+++ hivemq-edge/src/main/java/com/hivemq/api/resources/impl/ProtocolAdapterApiUtils.java
@@ -31,9 +31,8 @@ import com.hivemq.configuration.service.ConfigurationService;
 import com.hivemq.edge.HiveMQEdgeConstants;
 import com.hivemq.edge.VersionProvider;
 import com.hivemq.http.HttpConstants;
-import com.hivemq.protocols.ProtocolAdapterManager2;
+import com.hivemq.protocols.ProtocolAdapterManager;
 import com.hivemq.protocols.ProtocolAdapterSchemaManager;
-import com.hivemq.protocols.fsm.ProtocolAdapterManager2;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -66,7 +65,7 @@ public class ProtocolAdapterApiUtils {
      */
     public static @Nullable ProtocolAdapter convertInstalledAdapterType(
             final @NotNull ObjectMapper objectMapper,
-            final @NotNull ProtocolAdapterManager2 adapterManager,
+            final @NotNull ProtocolAdapterManager adapterManager,
             final @NotNull ProtocolAdapterInformation info,
             final @NotNull ConfigurationService configurationService,
             final @NotNull VersionProvider versionProvider,
