+++ hivemq-edge/src/main/java/com/hivemq/protocols/ProtocolAdapterManager.java
@@ -20,6 +20,7 @@ import static com.hivemq.persistence.domain.DomainTagAddResult.DomainTagPutStatu
 
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.base.Preconditions;
+import com.google.common.collect.Sets;
 import com.hivemq.adapter.sdk.api.ProtocolAdapter;
 import com.hivemq.adapter.sdk.api.ProtocolAdapterInformation;
 import com.hivemq.adapter.sdk.api.events.EventService;
@@ -27,8 +28,6 @@ import com.hivemq.adapter.sdk.api.events.model.Event;
 import com.hivemq.adapter.sdk.api.exceptions.ProtocolAdapterException;
 import com.hivemq.adapter.sdk.api.factories.ProtocolAdapterFactory;
 import com.hivemq.adapter.sdk.api.services.ProtocolAdapterMetricsService;
-import com.hivemq.adapter.sdk.api.services.ProtocolAdapterPublishService;
-import com.hivemq.adapter.sdk.api.state.ProtocolAdapterState;
 import com.hivemq.adapter.sdk.api.tag.Tag;
 import com.hivemq.configuration.entity.adapter.ProtocolAdapterEntity;
 import com.hivemq.configuration.reader.ProtocolAdapterExtractor;
@@ -43,20 +42,24 @@ import com.hivemq.edge.modules.adapters.metrics.ProtocolAdapterMetricsServiceImp
 import com.hivemq.edge.modules.api.adapters.ProtocolAdapterPollingService;
 import com.hivemq.persistence.domain.DomainTag;
 import com.hivemq.persistence.domain.DomainTagAddResult;
+import com.hivemq.protocols.fsm.I18nProtocolAdapterMessage;
+import com.hivemq.protocols.fsm.ProtocolAdapterManagerState;
 import com.hivemq.protocols.northbound.NorthboundConsumerFactory;
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -65,13 +68,32 @@ import org.jetbrains.annotations.VisibleForTesting;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-// NOTE: @Inject and @Singleton removed — replaced by ProtocolAdapterManager2 (FSM redesign Phase 6).
-// This class is kept for reference but is no longer wired into the DI graph.
-@SuppressWarnings({"unchecked", "FutureReturnValueIgnored"})
+/**
+ * Manages the lifecycle of all protocol adapter instances.
+ * <p>
+ * Responsibilities:
+ * <ol>
+ *   <li>Creates, starts, stops, and deletes {@link ProtocolAdapterWrapper} instances</li>
+ *   <li>Fires {@link Event}s on adapter lifecycle changes (start, stop, delete, config refresh)</li>
+ *   <li>Handles configuration refreshes: diffs old vs new config and creates/updates/deletes adapters accordingly</li>
+ *   <li>Delegates to {@link ProtocolAdapterFactoryManager} for adapter factory lookup</li>
+ *   <li>Provides backward-compatible API for consumers that previously used
+ *       {@link com.hivemq.protocols.ProtocolAdapterManager}</li>
+ * </ol>
+ * <p>
+ * Threading Model:
+ * <ul>
+ *   <li>Adapter map uses {@link ConcurrentHashMap} for thread-safe access</li>
+ *   <li>Configuration refresh runs on a single-threaded executor to serialize refresh operations</li>
+ *   <li>Within a refresh, adapter create/update/delete operations run sequentially on that refresh thread</li>
+ * </ul>
+ */
+@Singleton
 public class ProtocolAdapterManager {
-    private static final Logger log = LoggerFactory.getLogger(ProtocolAdapterManager.class);
-
-    private final @NotNull Map<String, ProtocolAdapterWrapper> protocolAdapters;
+    private static final String ADAPTER_ID = "adapterId";
+    private static final Logger LOGGER = LoggerFactory.getLogger(ProtocolAdapterManager.class);
+    // ConcurrentHashMap provides thread-safe access without explicit locking
+    private final @NotNull Map<String, ProtocolAdapterWrapper> protocolAdapterMap = new ConcurrentHashMap<>();
     private final @NotNull MetricRegistry metricRegistry;
     private final @NotNull HiveMQEdgeRemoteService remoteService;
     private final @NotNull EventService eventService;
@@ -85,8 +107,11 @@ public class ProtocolAdapterManager {
     private final @NotNull TagManager tagManager;
     private final @NotNull ProtocolAdapterExtractor protocolAdapterConfig;
     private final @NotNull ExecutorService executorService;
-    private final @NotNull ProtocolAdapterPublishService adapterPublishService;
+    private final @NotNull AtomicInteger refreshTasksInProgress = new AtomicInteger(0);
+    private final @NotNull AtomicReference<ProtocolAdapterManagerState> managerState =
+            new AtomicReference<>(ProtocolAdapterManagerState.Idle);
 
+    @Inject
     public ProtocolAdapterManager(
             final @NotNull MetricRegistry metricRegistry,
             final @NotNull HiveMQEdgeRemoteService remoteService,
@@ -113,7 +138,6 @@ public class ProtocolAdapterManager {
         this.northboundConsumerFactory = northboundConsumerFactory;
         this.tagManager = tagManager;
         this.protocolAdapterConfig = protocolAdapterConfig;
-        this.protocolAdapters = new ConcurrentHashMap<>();
         this.executorService = Executors.newSingleThreadExecutor();
         this.adapterPublishService = adapterPublishService;
         Runtime.getRuntime().addShutdownHook(new Thread(executorService::shutdown));
@@ -121,200 +145,303 @@ public class ProtocolAdapterManager {
                 protocolAdapterFactoryManager.writingEnabledChanged(protocolAdapterWritingService.writingEnabled()));
     }
 
-    // API, must be threadsafe
-
-    private static void syncFuture(final @NotNull Future<?> future) {
-        try {
-            future.get();
-        } catch (final InterruptedException e) {
-            Thread.currentThread().interrupt();
-            log.error("Interrupted while async execution: ", e.getCause());
-        } catch (final ExecutionException e) {
-            log.error("Exception happened while async execution: ", e.getCause());
-        }
+    /**
+     * Indicates whether one or more refresh operations are currently queued or running.
+     * <p>
+     * The value becomes {@code true} before a refresh task is handed to the executor and
+     * returns to {@code false} only after the refresh task finishes (successfully or with
+     * an error).
+     *
+     * @return {@code true} when refresh work is in progress, otherwise {@code false}
+     */
+    public boolean isBusy() {
+        return refreshTasksInProgress.get() > 0;
     }
 
-    public static <T> @NotNull T runWithContextLoader(
-            final @NotNull ClassLoader contextLoader, final @NotNull Supplier<T> wrapperSupplier) {
-        final ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
-        try {
-            Thread.currentThread().setContextClassLoader(contextLoader);
-            return wrapperSupplier.get();
-        } finally {
-            Thread.currentThread().setContextClassLoader(contextClassLoader);
-        }
+    /**
+     * Returns the high-level manager state.
+     *
+     * @return {@link ProtocolAdapterManagerState#Running} while refresh work is queued/running; otherwise {@link ProtocolAdapterManagerState#Idle}
+     */
+    public @NotNull ProtocolAdapterManagerState getState() {
+        return Objects.requireNonNull(managerState.get());
     }
 
+    /**
+     * Start the adapter manager — registers the config consumer to begin receiving adapter configs.
+     * This replaces the old {@code ProtocolAdapterManager.start()} method.
+     */
     public void start() {
-        if (log.isDebugEnabled()) {
-            log.debug("Starting adapters");
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Starting adapters");
         }
         protocolAdapterConfig.registerConsumer(this::refresh);
     }
 
+    /**
+     * Shutdown all running adapters.
+     */
     @VisibleForTesting
     public void shutdown() {
-        protocolAdapters.values().forEach(entry -> {
+        protocolAdapterMap.values().forEach(wrapper -> {
             try {
-                entry.stopAsync(true).get(5, TimeUnit.SECONDS);
-            } catch (final InterruptedException | ExecutionException | TimeoutException e) {
-                log.error("Exception happened while shutting down adapter: ", e);
+                wrapper.stop(true);
+            } catch (final Exception e) {
+                LOGGER.error("Exception happened while shutting down adapter '{}': ", wrapper.getAdapterId(), e);
             }
         });
     }
 
-    public void refresh(final @NotNull List<ProtocolAdapterEntity> configs) {
-        try {
-            executorService.submit(() -> {
-                log.info("Refreshing adapters");
+    // ===== Adapter Lookup =====
 
-                final Map<String, ProtocolAdapterConfig> protocolAdapterConfigs = configs.stream()
-                        .map(configConverter::fromEntity)
-                        .collect(Collectors.toMap(ProtocolAdapterConfig::getAdapterId, Function.identity()));
-
-                final List<String> loadListOfAdapterNames = new ArrayList<>(protocolAdapterConfigs.keySet());
-
-                final List<String> adaptersToBeDeleted = new ArrayList<>(protocolAdapters.keySet());
-                adaptersToBeDeleted.removeAll(loadListOfAdapterNames);
-
-                final List<String> adaptersToBeCreated = new ArrayList<>(loadListOfAdapterNames);
-                adaptersToBeCreated.removeAll(protocolAdapters.keySet());
-
-                final List<String> adaptersToBeUpdated = new ArrayList<>(protocolAdapters.keySet());
-                adaptersToBeUpdated.removeAll(adaptersToBeCreated);
-                adaptersToBeUpdated.removeAll(adaptersToBeDeleted);
-
-                final List<String> failedAdapters = new ArrayList<>();
-
-                adaptersToBeDeleted.forEach(name -> {
-                    try {
-                        if (log.isDebugEnabled()) {
-                            log.debug("Deleting adapter '{}'", name);
-                        }
-                        stopAsync(name, true)
-                                .whenComplete((ignored, t) -> deleteAdapterInternal(name))
-                                .get();
-                    } catch (final InterruptedException e) {
-                        Thread.currentThread().interrupt();
-                        failedAdapters.add(name);
-                        log.error("Interrupted while deleting adapter {}", name, e);
-                    } catch (final ExecutionException e) {
-                        failedAdapters.add(name);
-                        log.error("Failed deleting adapter {}", name, e);
-                    }
-                });
-
-                adaptersToBeCreated.forEach(name -> {
-                    try {
-                        if (log.isDebugEnabled()) {
-                            log.debug("Creating adapter '{}'", name);
-                        }
-                        final ProtocolAdapterConfig adapterConfig = protocolAdapterConfigs.get(name);
-                        if (adapterConfig == null) {
-                            log.error("Config for adapter '{}' not found, skipping creation", name);
-                            return;
-                        }
-                        startAsync(createAdapterInternal(adapterConfig, versionProvider.getVersion()))
-                                .get();
-                    } catch (final InterruptedException e) {
-                        Thread.currentThread().interrupt();
-                        failedAdapters.add(name);
-                        log.error("Interrupted while adding adapter {}", name, e);
-                    } catch (final ExecutionException e) {
-                        failedAdapters.add(name);
-                        log.error("Failed adding adapter {}", name, e);
-                    }
-                });
-
-                adaptersToBeUpdated.forEach(name -> {
-                    try {
-                        final var wrapper = protocolAdapters.get(name);
-                        if (wrapper == null) {
-                            log.error(
-                                    "Existing adapters were modified while a refresh was ongoing, adapter with name '{}' was deleted and could not be updated",
-                                    name);
-                        }
-                        final ProtocolAdapterConfig updatedConfig = protocolAdapterConfigs.get(name);
-                        if (updatedConfig == null) {
-                            log.error("Config for adapter '{}' not found, skipping update", name);
-                            return;
-                        }
-                        if (wrapper != null && !updatedConfig.equals(wrapper.getConfig())) {
-                            if (log.isDebugEnabled()) {
-                                log.debug("Updating adapter '{}'", name);
-                            }
-                            stopAsync(name, true)
-                                    .thenApply(v -> {
-                                        deleteAdapterInternal(name);
-                                        return null;
-                                    })
-                                    .thenCompose(ignored -> startAsync(
-                                            createAdapterInternal(updatedConfig, versionProvider.getVersion())))
-                                    .get();
-                        } else {
-                            if (log.isDebugEnabled()) {
-                                log.debug("Not-updating adapter '{}' since the config is unchanged", name);
-                            }
-                        }
-                    } catch (final InterruptedException e) {
-                        Thread.currentThread().interrupt();
-                        failedAdapters.add(name);
-                        log.error("Interrupted while updating adapter {}", name, e);
-                    } catch (final ExecutionException e) {
-                        failedAdapters.add(name);
-                        log.error("Failed updating adapter {}", name, e);
-                    }
-                });
-
-                if (failedAdapters.isEmpty()) {
-                    eventService
-                            .configurationEvent()
-                            .withSeverity(Event.SEVERITY.INFO)
-                            .withMessage("Configuration has been successfully updated")
-                            .fire();
-                } else {
-                    eventService
-                            .configurationEvent()
-                            .withSeverity(Event.SEVERITY.CRITICAL)
-                            .withMessage("Reloading of configuration failed")
-                            .fire();
-                }
-            });
-        } catch (final RejectedExecutionException e) {
-            if (executorService.isShutdown()) {
-                log.debug("Adapter refresh task submission rejected, executor is shutting down.");
-            } else {
-                throw e;
-            }
-        }
+    public @NotNull Optional<ProtocolAdapterWrapper> getProtocolAdapterWrapperByAdapterId(
+            final @NotNull String adapterId) {
+        Preconditions.checkNotNull(adapterId);
+        return Optional.ofNullable(protocolAdapterMap.get(adapterId));
     }
 
-    // INTERNAL
+    protected @NotNull Set<String> getProtocolAdapterIdSet() {
+        return new HashSet<>(protocolAdapterMap.keySet());
+    }
+
+    /**
+     * Returns a defensive copy of all protocol adapters.
+     * Used by REST API endpoints that need to iterate over all adapters.
+     *
+     * @return an unmodifiable copy of the adapter map
+     */
+    public @NotNull Map<String, ProtocolAdapterWrapper> getProtocolAdapters() {
+        return Map.copyOf(protocolAdapterMap);
+    }
+
+    // ===== Factory/Type Lookups (delegate to ProtocolAdapterFactoryManager) =====
 
     public boolean protocolAdapterFactoryExists(final @NotNull String protocolAdapterType) {
         Preconditions.checkNotNull(protocolAdapterType);
         return protocolAdapterFactoryManager.get(protocolAdapterType).isPresent();
     }
 
+    public @NotNull Optional<ProtocolAdapterInformation> getAdapterTypeById(final @NotNull String typeId) {
+        Preconditions.checkNotNull(typeId);
+        return Optional.ofNullable(getAllAvailableAdapterTypes().get(typeId));
+    }
+
+    public @NotNull Map<String, ProtocolAdapterInformation> getAllAvailableAdapterTypes() {
+        return protocolAdapterFactoryManager.getAllAvailableAdapterTypes();
+    }
+
+    public boolean writingEnabled() {
+        return protocolAdapterWritingService.writingEnabled();
+    }
+
+    // ===== Domain Tag Operations =====
+
+    public @NotNull DomainTagAddResult addDomainTag(
+            final @NotNull String adapterId, final @NotNull DomainTag domainTag) {
+        return getProtocolAdapterWrapperByAdapterId(adapterId)
+                .map(wrapper -> {
+                    final var tags = new ArrayList<>(wrapper.getTags());
+                    final boolean alreadyExists =
+                            tags.stream().anyMatch(t -> t.getName().equals(domainTag.getTagName()));
+                    if (!alreadyExists) {
+                        tags.add(configConverter.domainTagToTag(
+                                wrapper.getProtocolAdapterInformation().getProtocolId(), domainTag));
+                        updateAdapterTags(adapterId, tags);
+                        return DomainTagAddResult.success();
+                    } else {
+                        return DomainTagAddResult.failed(ALREADY_EXISTS, adapterId);
+                    }
+                })
+                .orElse(DomainTagAddResult.failed(ADAPTER_MISSING, adapterId));
+    }
+
+    public @NotNull List<DomainTag> getDomainTags() {
+        return protocolAdapterMap.values().stream()
+                .flatMap(wrapper -> wrapper.getTags().stream()
+                        .map(tag -> new DomainTag(
+                                tag.getName(),
+                                wrapper.getId(),
+                                tag.getDescription(),
+                                configConverter.convertTagDefinitionToJsonNode(tag.getDefinition()))))
+                .toList();
+    }
+
+    public @NotNull Optional<DomainTag> getDomainTagByName(final @NotNull String tagName) {
+        return protocolAdapterMap.values().stream()
+                .flatMap(wrapper -> wrapper.getTags().stream()
+                        .filter(t -> t.getName().equals(tagName))
+                        .map(tag -> new DomainTag(
+                                tag.getName(),
+                                wrapper.getId(),
+                                tag.getDescription(),
+                                configConverter.convertTagDefinitionToJsonNode(tag.getDefinition()))))
+                .findFirst();
+    }
+
+    public @NotNull Optional<List<DomainTag>> getTagsForAdapter(final @NotNull String adapterId) {
+        return getProtocolAdapterWrapperByAdapterId(adapterId).map(wrapper -> wrapper.getTags().stream()
+                .map(tag -> new DomainTag(
+                        tag.getName(),
+                        wrapper.getId(),
+                        tag.getDescription(),
+                        configConverter.convertTagDefinitionToJsonNode(tag.getDefinition())))
+                .toList());
+    }
+
+    private boolean updateAdapterTags(final @NotNull String adapterId, final @NotNull List<? extends Tag> tags) {
+        Preconditions.checkNotNull(adapterId);
+        return getProtocolAdapterWrapperByAdapterId(adapterId)
+                .map(wrapper -> {
+                    final var protocolId = wrapper.getAdapterInformation().getProtocolId();
+                    final var newConfig = new ProtocolAdapterConfig(
+                            wrapper.getId(),
+                            protocolId,
+                            wrapper.getAdapterInformation().getCurrentConfigVersion(),
+                            wrapper.getConfigObject(),
+                            wrapper.getSouthboundMappings(),
+                            wrapper.getNorthboundMappings(),
+                            tags);
+                    updateAdapter(newConfig);
+                    return true;
+                })
+                .orElse(false);
+    }
+
+    private void updateAdapter(final @NotNull ProtocolAdapterConfig config) {
+        deleteProtocolAdapterByAdapterId(config.getAdapterId());
+        createProtocolAdapter(config, versionProvider.getVersion());
+        try {
+            start(config.getAdapterId());
+        } catch (final ProtocolAdapterException e) {
+            LOGGER.error("Failed to start adapter '{}' after update", config.getAdapterId(), e);
+        }
+    }
+
+    // ===== Synchronous Start/Stop =====
+
+    /**
+     * Start an adapter by ID.
+     * <p>
+     * Delegates to {@link ProtocolAdapterWrapper#start()} and fires an adapter event
+     * indicating success or failure.
+     *
+     * @param adapterId the adapter to start
+     * @throws ProtocolAdapterException if the adapter is not found or fails to start
+     */
+    public void start(final @NotNull String adapterId) throws ProtocolAdapterException {
+        LOGGER.info("Starting protocol-adapter '{}'.", adapterId);
+        final var optionalWrapper = getProtocolAdapterWrapperByAdapterId(adapterId);
+        if (optionalWrapper.isEmpty()) {
+            throw new ProtocolAdapterException(
+                    I18nProtocolAdapterMessage.PROTOCOL_ADAPTER_MANAGER_PROTOCOL_ADAPTER_NOT_FOUND.get(
+                            Map.of(ADAPTER_ID, adapterId)));
+        }
+        final ProtocolAdapterWrapper wrapper = optionalWrapper.get();
+        final String protocolId = wrapper.getProtocolAdapterInformation().getProtocolId();
+        final boolean success = wrapper.start();
+
+        if (success) {
+            LOGGER.info("Protocol-adapter '{}' started successfully.", adapterId);
+            final HiveMQEdgeRemoteEvent event =
+                    new HiveMQEdgeRemoteEvent(HiveMQEdgeRemoteEvent.EVENT_TYPE.ADAPTER_STARTED);
+            event.addUserData("adapterType", protocolId);
+            remoteService.fireUsageEvent(event);
+            eventService
+                    .createAdapterEvent(adapterId, protocolId)
+                    .withSeverity(Event.SEVERITY.INFO)
+                    .withMessage("Adapter '" + adapterId + "' started OK.")
+                    .fire();
+        } else {
+            LOGGER.warn("Protocol-adapter '{}' could not be started, reason: {}", adapterId, "unknown");
+            final HiveMQEdgeRemoteEvent event =
+                    new HiveMQEdgeRemoteEvent(HiveMQEdgeRemoteEvent.EVENT_TYPE.ADAPTER_ERROR);
+            event.addUserData("adapterType", protocolId);
+            remoteService.fireUsageEvent(event);
+            eventService
+                    .createAdapterEvent(adapterId, protocolId)
+                    .withSeverity(Event.SEVERITY.CRITICAL)
+                    .withMessage("Error starting adapter '" + adapterId + "'.")
+                    .fire();
+            throw new ProtocolAdapterException("Failed to start adapter: " + adapterId);
+        }
+    }
+
+    /**
+     * Stop an adapter by ID.
+     *
+     * @param adapterId the adapter to stop
+     * @param destroy   whether to call {@link ProtocolAdapter#destroy()} after stopping
+     * @throws ProtocolAdapterException if the adapter is not found
+     */
+    public void stop(final @NotNull String adapterId, final boolean destroy) throws ProtocolAdapterException {
+        LOGGER.info("Stopping protocol-adapter '{}'.", adapterId);
+        final var optionalWrapper = getProtocolAdapterWrapperByAdapterId(adapterId);
+        if (optionalWrapper.isEmpty()) {
+            throw new ProtocolAdapterException(
+                    I18nProtocolAdapterMessage.PROTOCOL_ADAPTER_MANAGER_PROTOCOL_ADAPTER_NOT_FOUND.get(
+                            Map.of(ADAPTER_ID, adapterId)));
+        }
+        final ProtocolAdapterWrapper wrapper = optionalWrapper.get();
+        final String protocolId = wrapper.getProtocolAdapterInformation().getProtocolId();
+        final boolean success = wrapper.stop(destroy);
+
+        if (success) {
+            LOGGER.info("Protocol-adapter '{}' stopped successfully.", adapterId);
+            eventService
+                    .createAdapterEvent(adapterId, protocolId)
+                    .withSeverity(Event.SEVERITY.INFO)
+                    .withMessage("Adapter '" + adapterId + "' stopped OK.")
+                    .fire();
+        } else {
+            LOGGER.warn("Protocol-adapter '{}' was unable to stop cleanly", adapterId);
+            eventService
+                    .createAdapterEvent(adapterId, protocolId)
+                    .withSeverity(Event.SEVERITY.CRITICAL)
+                    .withMessage("Error stopping adapter '" + adapterId + "'.")
+                    .fire();
+        }
+    }
+
+    // ===== Async Start/Stop (for REST API backward compatibility) =====
+
+    /**
+     * Async wrapper for {@link #start(String)} — matches the old manager's API.
+     */
     public @NotNull CompletableFuture<Void> startAsync(final @NotNull String protocolAdapterId) {
         Preconditions.checkNotNull(protocolAdapterId);
-        return getProtocolAdapterWrapperByAdapterId(protocolAdapterId)
-                .map(this::startAsync)
-                .orElseGet(() -> CompletableFuture.failedFuture(
-                        new ProtocolAdapterException("Adapter '" + protocolAdapterId + "'not found.")));
+        if (!protocolAdapterMap.containsKey(protocolAdapterId)) {
+            return CompletableFuture.failedFuture(
+                    new ProtocolAdapterException("Adapter not found: " + protocolAdapterId));
+        }
+        return CompletableFuture.runAsync(() -> {
+            try {
+                start(protocolAdapterId);
+            } catch (final ProtocolAdapterException e) {
+                throw new RuntimeException(e);
+            }
+        });
     }
 
+    /**
+     * Async wrapper for {@link #stop(String, boolean)} — matches the old manager's API.
+     */
     public @NotNull CompletableFuture<Void> stopAsync(final @NotNull String protocolAdapterId, final boolean destroy) {
         Preconditions.checkNotNull(protocolAdapterId);
-        return getProtocolAdapterWrapperByAdapterId(protocolAdapterId)
-                .map(wrapper -> stopAsync(wrapper, destroy))
-                .orElseGet(() -> CompletableFuture.failedFuture(
-                        new ProtocolAdapterException("Adapter '" + protocolAdapterId + "'not found.")));
+        return CompletableFuture.runAsync(() -> {
+            try {
+                stop(protocolAdapterId, destroy);
+            } catch (final ProtocolAdapterException e) {
+                throw new RuntimeException(e);
+            }
+        });
     }
 
-    private @NotNull ProtocolAdapterWrapper createAdapterInternal(
-            final @NotNull ProtocolAdapterConfig config, final @NotNull String version) {
-        return protocolAdapters.computeIfAbsent(config.getAdapterId(), ignored -> {
+    // ===== Adapter Creation/Deletion =====
+
+    protected void createProtocolAdapter(final @NotNull ProtocolAdapterConfig config, final @NotNull String version) {
+        Preconditions.checkNotNull(config);
+        final String adapterId = config.getAdapterId();
+        protocolAdapterMap.computeIfAbsent(adapterId, ignored -> {
             final String configProtocolId = config.getProtocolId();
             // legacy handling, hardcoded here, to not add legacy stuff into the adapter-sdk
             final String adapterType =
@@ -331,28 +458,22 @@ public class ProtocolAdapterManager {
             }
             final ProtocolAdapterFactory<?> factory = maybeFactory.get();
 
-            log.info("Found configuration for adapter {} / {}", config.getAdapterId(), adapterType);
-            config.missingTags().ifPresent(missing -> {
+            LOGGER.info("Found configuration for adapter {} / {}", config.getAdapterId(), adapterType);
+            config.missingTags().ifPresent(missingTag -> {
                 throw new IllegalArgumentException(
-                        "Tags used in mappings but not configured in adapter " + config.getAdapterId()
-                                + ": "
-                                + missing);
+                        "Tags used in mappings but not configured in adapter " + adapterType + ": " + missingTag);
             });
 
             return runWithContextLoader(factory.getClass().getClassLoader(), () -> {
                 final ProtocolAdapterMetricsService metricsService =
                         new ProtocolAdapterMetricsServiceImpl(configProtocolId, config.getAdapterId(), metricRegistry);
-
-                final ProtocolAdapterStateImpl state =
-                        new ProtocolAdapterStateImpl(eventService, config.getAdapterId(), configProtocolId);
-
+                final ProtocolAdapterStateImpl state = new ProtocolAdapterStateImpl(
+                        moduleServices.eventService(), config.getAdapterId(), configProtocolId);
                 final ModuleServicesPerModuleImpl perModule = new ModuleServicesPerModuleImpl(
                         adapterPublishService,
                         eventService,
                         protocolAdapterWritingService,
-                        new ProtocolAdapterTagStreamingServiceImpl(
-                                config.getAdapterId(), tagManager, dataPointBuilder -> {}));
-
+                        tagManager);
                 final ProtocolAdapter protocolAdapter = factory.createAdapter(
                         factory.getInformation(),
                         new ProtocolAdapterInputImpl(
@@ -366,209 +487,199 @@ public class ProtocolAdapterManager {
                                 metricsService));
                 // hen-egg problem. Rather solve this here as have not final fields in the adapter.
                 perModule.setAdapter(protocolAdapter);
-
-                protocolAdapterMetrics.increaseProtocolAdapterMetric(configProtocolId);
-                return new ProtocolAdapterWrapper(
-                        metricsService,
-                        protocolAdapterWritingService,
-                        protocolAdapterPollingService,
-                        config,
+                final ProtocolAdapterWrapper wrapper = new ProtocolAdapterWrapper(
                         protocolAdapter,
+                        config,
                         factory,
                         factory.getInformation(),
+                        metricsService,
                         state,
-                        northboundConsumerFactory,
+                        protocolAdapterPollingService,
+                        eventService,
+                        perModule,
                         tagManager,
-                        perModule);
+                        northboundConsumerFactory,
+                        protocolAdapterWritingService);
+                protocolAdapterMetrics.increaseProtocolAdapterMetric(configProtocolId);
+                return wrapper;
             });
         });
     }
 
-    private void deleteAdapterInternal(final @NotNull String adapterId) {
-        final ProtocolAdapterWrapper wrapper = protocolAdapters.remove(adapterId);
-        if (wrapper != null) {
-            protocolAdapterMetrics.decreaseProtocolAdapterMetric(
-                    wrapper.getAdapterInformation().getProtocolId());
-            eventService
-                    .createAdapterEvent(
-                            adapterId, wrapper.getProtocolAdapterInformation().getProtocolId())
-                    .withSeverity(Event.SEVERITY.WARN)
-                    .withMessage("Adapter '" + adapterId + "' was deleted from the system permanently.")
-                    .fire();
-        } else {
-            log.warn("Tried to delete adapter '{}' but it was not found in the system.", adapterId);
+    protected @NotNull Optional<ProtocolAdapterWrapper> deleteProtocolAdapterWrapperByAdapterId(
+            final @NotNull String adapterId) {
+        Preconditions.checkNotNull(adapterId);
+        return Optional.ofNullable(protocolAdapterMap.remove(adapterId));
+    }
+
+    protected void deleteProtocolAdapterByAdapterId(final @NotNull String adapterId) {
+        deleteProtocolAdapterWrapperByAdapterId(adapterId)
+                .ifPresentOrElse(
+                        wrapper -> {
+                            final String protocolId =
+                                    wrapper.getProtocolAdapterInformation().getProtocolId();
+                            protocolAdapterMetrics.decreaseProtocolAdapterMetric(protocolId);
+                            eventService
+                                    .createAdapterEvent(adapterId, protocolId)
+                                    .withSeverity(Event.SEVERITY.WARN)
+                                    .withMessage("Adapter '" + adapterId + "' was deleted from the system permanently.")
+                                    .fire();
+                        },
+                        () -> LOGGER.warn(
+                                "Tried to delete adapter '{}' but it was not found in the system.", adapterId));
+    }
+
+    // ===== Configuration Refresh =====
+
+    /**
+     * Refreshes protocol adapters from the latest configuration snapshot.
+     * <p>
+     * The operation is serialized on a dedicated single-thread executor. Any concurrent calls are queued
+     * and processed in order.
+     *
+     * @param configs the latest adapter configuration entities
+     */
+    public void refresh(final @NotNull List<ProtocolAdapterEntity> configs) {
+        refreshTasksInProgress.incrementAndGet();
+        managerState.set(ProtocolAdapterManagerState.Running);
+        try {
+            executorService.execute(() -> {
+                LOGGER.info("Refreshing adapters");
+                try {
+                    final Map<String, ProtocolAdapterConfig> protocolAdapterConfigs = configs.stream()
+                            .map(configConverter::fromEntity)
+                            .collect(Collectors.toMap(ProtocolAdapterConfig::getAdapterId, Function.identity()));
+
+                    final Set<String> oldProtocolAdapterIdSet = getProtocolAdapterIdSet();
+                    final Set<String> newProtocolAdapterIdSet = new HashSet<>(protocolAdapterConfigs.keySet());
+
+                    final Set<String> toBeDeletedProtocolAdapterIdSet =
+                            new HashSet<>(Sets.difference(oldProtocolAdapterIdSet, newProtocolAdapterIdSet));
+                    final Set<String> toBeCreatedProtocolAdapterIdSet =
+                            new HashSet<>(Sets.difference(newProtocolAdapterIdSet, oldProtocolAdapterIdSet));
+                    final Set<String> toBeUpdatedProtocolAdapterIdSet =
+                            new HashSet<>(Sets.intersection(newProtocolAdapterIdSet, oldProtocolAdapterIdSet));
+
+                    final Set<String> failedAdapterSet = new HashSet<>();
+                    refreshDeletedAdapters(toBeDeletedProtocolAdapterIdSet, failedAdapterSet);
+                    refreshCreatedAdapters(toBeCreatedProtocolAdapterIdSet, protocolAdapterConfigs, failedAdapterSet);
+                    refreshUpdatedAdapters(toBeUpdatedProtocolAdapterIdSet, protocolAdapterConfigs, failedAdapterSet);
+
+                    if (failedAdapterSet.isEmpty()) {
+                        eventService
+                                .configurationEvent()
+                                .withSeverity(Event.SEVERITY.INFO)
+                                .withMessage("Configuration has been successfully updated")
+                                .fire();
+                    } else {
+                        eventService
+                                .configurationEvent()
+                                .withSeverity(Event.SEVERITY.CRITICAL)
+                                .withMessage("Reloading of configuration failed")
+                                .fire();
+                    }
+                } catch (final Exception e) {
+                    LOGGER.error("Failed refreshing adapters", e);
+                } finally {
+                    final int remainingTasks = refreshTasksInProgress.decrementAndGet();
+                    if (remainingTasks == 0) {
+                        managerState.set(ProtocolAdapterManagerState.Idle);
+                    }
+                }
+            });
+        } catch (final RuntimeException e) {
+            final int remainingTasks = refreshTasksInProgress.decrementAndGet();
+            if (remainingTasks == 0) {
+                managerState.set(ProtocolAdapterManagerState.Idle);
+            }
+            throw e;
         }
     }
 
-    @VisibleForTesting
-    @NotNull
-    CompletableFuture<Void> startAsync(final @NotNull ProtocolAdapterWrapper wrapper) {
-        Preconditions.checkNotNull(wrapper);
-        final String wid = wrapper.getId();
-        log.info("Starting protocol-adapter '{}'.", wid);
-        return wrapper.startAsync(writingEnabled()).whenComplete((result, throwable) -> {
-            if (throwable == null) {
-                log.info("Protocol-adapter '{}' started successfully.", wid);
-                fireEvent(
-                        wrapper,
-                        HiveMQEdgeRemoteEvent.EVENT_TYPE.ADAPTER_STARTED,
-                        Event.SEVERITY.INFO,
-                        "Adapter '" + wid + "' started OK.");
-            } else {
-                log.warn("Protocol-adapter '{}' could not be started, reason: {}", wid, "unknown");
-                fireEvent(
-                        wrapper,
-                        HiveMQEdgeRemoteEvent.EVENT_TYPE.ADAPTER_ERROR,
-                        Event.SEVERITY.CRITICAL,
-                        "Error starting adapter '" + wid + "'.");
+    private void refreshDeletedAdapters(
+            final @NotNull Set<String> adapterIds, final @NotNull Set<String> failedAdapterSet) {
+        for (final String adapterId : adapterIds) {
+            try {
+                if (LOGGER.isDebugEnabled()) {
+                    LOGGER.debug("Deleting adapter '{}'", adapterId);
+                }
+                stop(adapterId, true);
+                deleteProtocolAdapterByAdapterId(adapterId);
+            } catch (final Exception e) {
+                failedAdapterSet.add(adapterId);
+                LOGGER.error("Failed deleting adapter {}", adapterId, e);
             }
-        });
+        }
     }
 
-    private void fireEvent(
-            final @NotNull ProtocolAdapterWrapper wrapper,
-            final @NotNull HiveMQEdgeRemoteEvent.EVENT_TYPE eventType,
-            final @NotNull Event.SEVERITY severity,
-            final @NotNull String message) {
-        final String protocolId = wrapper.getProtocolAdapterInformation().getProtocolId();
-        final HiveMQEdgeRemoteEvent event = new HiveMQEdgeRemoteEvent(eventType);
-        event.addUserData("adapterType", protocolId);
-        remoteService.fireUsageEvent(event);
-        eventService
-                .createAdapterEvent(wrapper.getId(), protocolId)
-                .withSeverity(severity)
-                .withMessage(message)
-                .fire();
-    }
-
-    @VisibleForTesting
-    @NotNull
-    CompletableFuture<Void> stopAsync(final @NotNull ProtocolAdapterWrapper wrapper, final boolean destroy) {
-        Preconditions.checkNotNull(wrapper);
-        log.info("Stopping protocol-adapter '{}'.", wrapper.getId());
-
-        return wrapper.stopAsync(destroy).whenComplete((result, throwable) -> {
-            wrapper.setRuntimeStatus(ProtocolAdapterState.RuntimeStatus.STOPPED);
-            final Event.SEVERITY severity;
-            final String message;
-            final String wid = wrapper.getId();
-            final String protocolId = wrapper.getProtocolAdapterInformation().getProtocolId();
-            if (throwable == null) {
-                log.info("Protocol-adapter '{}' stopped successfully.", wid);
-                severity = Event.SEVERITY.INFO;
-                message = "Adapter '" + wid + "' stopped OK.";
-            } else {
-                log.warn("Protocol-adapter '{}' was unable to stop cleanly", wrapper.getId());
-                severity = Event.SEVERITY.CRITICAL;
-                message = "Error stopping adapter '" + wid + "'.";
+    private void refreshCreatedAdapters(
+            final @NotNull Set<String> adapterIds,
+            final @NotNull Map<String, ProtocolAdapterConfig> protocolAdapterConfigs,
+            final @NotNull Set<String> failedAdapterSet) {
+        for (final String adapterId : adapterIds) {
+            try {
+                if (LOGGER.isDebugEnabled()) {
+                    LOGGER.debug("Creating adapter '{}'", adapterId);
+                }
+                createProtocolAdapter(
+                        Objects.requireNonNull(protocolAdapterConfigs.get(adapterId)), versionProvider.getVersion());
+                start(adapterId);
+            } catch (final Exception e) {
+                failedAdapterSet.add(adapterId);
+                LOGGER.error("Failed creating adapter {}", adapterId, e);
             }
-            eventService
-                    .createAdapterEvent(wid, protocolId)
-                    .withSeverity(severity)
-                    .withMessage(message)
-                    .fire();
-        });
+        }
     }
 
-    private void updateAdapter(final ProtocolAdapterConfig protocolAdapterConfig) {
-        deleteAdapterInternal(protocolAdapterConfig.getAdapterId());
-        syncFuture(startAsync(createAdapterInternal(protocolAdapterConfig, versionProvider.getVersion())));
-    }
-
-    private boolean updateAdapterTags(final @NotNull String adapterId, final @NotNull List<? extends Tag> tags) {
-        Preconditions.checkNotNull(adapterId);
-        return getProtocolAdapterWrapperByAdapterId(adapterId)
-                .map(wrapper -> {
-                    final var protocolId = wrapper.getAdapterInformation().getProtocolId();
-                    final var protocolAdapterConfig = new ProtocolAdapterConfig(
-                            wrapper.getId(),
-                            protocolId,
-                            wrapper.getAdapterInformation().getCurrentConfigVersion(),
-                            wrapper.getConfigObject(),
-                            wrapper.getSouthboundMappings(),
-                            wrapper.getNorthboundMappings(),
-                            tags);
-                    updateAdapter(protocolAdapterConfig);
-                    return true;
-                })
-                .orElse(false);
-    }
-
-    public @NotNull Optional<ProtocolAdapterWrapper> getProtocolAdapterWrapperByAdapterId(final @NotNull String id) {
-        Preconditions.checkNotNull(id);
-        return Optional.ofNullable(protocolAdapters.get(id));
-    }
-
-    public @NotNull Optional<ProtocolAdapterInformation> getAdapterTypeById(final @NotNull String typeId) {
-        Preconditions.checkNotNull(typeId);
-        final ProtocolAdapterInformation information =
-                getAllAvailableAdapterTypes().get(typeId);
-        return Optional.ofNullable(information);
-    }
-
-    public @NotNull Map<String, ProtocolAdapterInformation> getAllAvailableAdapterTypes() {
-        return protocolAdapterFactoryManager.getAllAvailableAdapterTypes();
-    }
-
-    public @NotNull Map<String, ProtocolAdapterWrapper> getProtocolAdapters() {
-        return Map.copyOf(protocolAdapters);
-    }
-
-    public boolean writingEnabled() {
-        return protocolAdapterWritingService.writingEnabled();
-    }
-
-    public @NotNull DomainTagAddResult addDomainTag(
-            final @NotNull String adapterId, final @NotNull DomainTag domainTag) {
-        return getProtocolAdapterWrapperByAdapterId(adapterId)
-                .map(wrapper -> {
-                    final var tags = new ArrayList<>(wrapper.getTags());
-                    final boolean alreadyExists =
-                            tags.stream().anyMatch(t -> t.getName().equals(domainTag.getTagName()));
-                    if (!alreadyExists) {
-                        tags.add(configConverter.domainTagToTag(
-                                wrapper.getProtocolAdapterInformation().getProtocolId(), domainTag));
-
-                        updateAdapterTags(adapterId, tags);
-                        return DomainTagAddResult.success();
-                    } else {
-                        return DomainTagAddResult.failed(ALREADY_EXISTS, adapterId);
+    private void refreshUpdatedAdapters(
+            final @NotNull Set<String> adapterIds,
+            final @NotNull Map<String, ProtocolAdapterConfig> protocolAdapterConfigs,
+            final @NotNull Set<String> failedAdapterSet) {
+        for (final String adapterId : adapterIds) {
+            try {
+                final ProtocolAdapterWrapper wrapper = protocolAdapterMap.get(adapterId);
+                if (wrapper == null) {
+                    failedAdapterSet.add(adapterId);
+                    LOGGER.error(
+                            "Existing adapters were modified while a refresh was ongoing, adapter '{}' was deleted and could not be updated",
+                            adapterId);
+                    continue;
+                }
+                if (Objects.requireNonNull(protocolAdapterConfigs.get(adapterId))
+                        .equals(wrapper.getConfig())) {
+                    if (LOGGER.isDebugEnabled()) {
+                        LOGGER.debug("Not-updating adapter '{}' since the config is unchanged", adapterId);
                     }
-                })
-                .orElse(DomainTagAddResult.failed(ADAPTER_MISSING, adapterId));
+                    continue;
+                }
+                if (LOGGER.isDebugEnabled()) {
+                    LOGGER.debug("Updating adapter '{}'", adapterId);
+                }
+                stop(adapterId, true);
+                deleteProtocolAdapterByAdapterId(adapterId);
+                createProtocolAdapter(protocolAdapterConfigs.get(adapterId), versionProvider.getVersion());
+                start(adapterId);
+            } catch (final Exception e) {
+                failedAdapterSet.add(adapterId);
+                LOGGER.error("Failed updating adapter {}", adapterId, e);
+            }
+        }
     }
 
-    public @NotNull List<DomainTag> getDomainTags() {
-        return protocolAdapters.values().stream()
-                .flatMap(wrapper -> wrapper.getTags().stream()
-                        .map(tag -> new DomainTag(
-                                tag.getName(),
-                                wrapper.getId(),
-                                tag.getDescription(),
-                                configConverter.convertTagDefinitionToJsonNode(tag.getDefinition()))))
-                .toList();
-    }
+    // ===== Static Utility (backward compat) =====
 
-    public @NotNull Optional<DomainTag> getDomainTagByName(final @NotNull String tagName) {
-        return protocolAdapters.values().stream()
-                .flatMap(wrapper -> wrapper.getTags().stream()
-                        .filter(t -> t.getName().equals(tagName))
-                        .map(tag -> new DomainTag(
-                                tag.getName(),
-                                wrapper.getId(),
-                                tag.getDescription(),
-                                configConverter.convertTagDefinitionToJsonNode(tag.getDefinition()))))
-                .findFirst();
-    }
-
-    public @NotNull Optional<List<DomainTag>> getTagsForAdapter(final @NotNull String adapterId) {
-        return getProtocolAdapterWrapperByAdapterId(adapterId).map(adapterToConfig -> adapterToConfig.getTags().stream()
-                .map(tag -> new DomainTag(
-                        tag.getName(),
-                        adapterToConfig.getId(),
-                        tag.getDescription(),
-                        configConverter.convertTagDefinitionToJsonNode(tag.getDefinition())))
-                .toList());
+    /**
+     * Runs a supplier with a specific classloader set as the thread's context classloader.
+     */
+    public static <T> @NotNull T runWithContextLoader(
+            final @NotNull ClassLoader contextLoader, final @NotNull Supplier<T> supplier) {
+        final ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
+        try {
+            Thread.currentThread().setContextClassLoader(contextLoader);
+            return supplier.get();
+        } finally {
+            Thread.currentThread().setContextClassLoader(contextClassLoader);
+        }
     }
 
     @VisibleForTesting
